Working ai agent with validation issue fix and ux improvement
