--- AI_Chat_Agent/repo/src/utils/constants.js
@@ -1,14 +1,0 @@
-// Configuration defaults - Updated with latest models
-export const DEFAULT_CONFIG = {
-  aiProvider: 'anthropic',
-  navigatorModel: 'claude-3-5-sonnet-20241022',
-  plannerModel: 'claude-3-5-sonnet-20241022', 
-  validatorModel: 'claude-3-haiku-20240307',
-  autoLogin: false,
-  safeMode: true,
-  voiceInput: true,
-  humanDelay: true,
-  debugMode: false,
-  maxRetries: 3,
-  timeout: 30000
-};
