+++ Social Shopping Agent/repo/.gitignore
@@ -10,6 +10,7 @@
 
 # production
 /build
+.env
 
 # misc
 .DS_Store
