+++ Social Shopping Agent/repo/package.json
