+++ Social Shopping Agent/repo/public/manifest.json
