+++ Social Shopping Agent/repo/src/components/AuthPage.js
