+++ Social Shopping Agent/repo/src/components/ChatInput.js
