+++ Social Shopping Agent/repo/src/components/ProfilePage.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { 
   FaUser, 
@@ -12,14 +12,45 @@ import {
   FaInfinity,
   FaClock,
   FaStar,
-  FaCog
+  FaCog,
+  FaToggleOn,
+  FaToggleOff
 } from 'react-icons/fa';
 
 const ProfilePage = ({ user, subscription, onLogout }) => {
   const navigate = useNavigate();
+  const [isToggling, setIsToggling] = useState(false);
 
   const handleBack = () => {
     navigate('/chat');
+  };
+
+  const handleTogglePersonalAPI = async () => {
+    if (isToggling) return;
+    
+    setIsToggling(true);
+    
+    try {
+      const newPreference = !subscription.userPreferPersonalAPI;
+      
+      // If turning ON and no API keys, redirect to settings
+      if (newPreference && !subscription.hasPersonalKeys) {
+        navigate('/settings');
+        setIsToggling(false);
+        return;
+      }
+      
+      // Set the preference
+      const success = await subscription.setUserAPIPreference(newPreference);
+      
+      if (success) {
+        console.log(`API preference set to: ${newPreference ? 'Personal API' : 'Free Trial'}`);
+      }
+    } catch (error) {
+      console.error('Error toggling API preference:', error);
+    } finally {
+      setIsToggling(false);
+    }
   };
 
   const formatDate = (dateString) => {
@@ -104,37 +135,35 @@ const ProfilePage = ({ user, subscription, onLogout }) => {
   const contentStyle = {
     flex: 1,
     overflowY: 'auto',
-    padding: '0',
-    WebkitOverflowScrolling: 'touch'
+    padding: '16px',
+    scrollbarWidth: 'none',
+    msOverflowStyle: 'none'
   };
 
   const sectionStyle = {
-    padding: '16px',
-    borderBottom: '1px solid rgba(255, 220, 220, 0.2)'
+    marginBottom: '24px'
   };
 
   const cardStyle = {
-    backgroundColor: '#003A7CFF',
+    backgroundColor: 'rgba(255, 220, 220, 0.08)',
+    border: '1px solid rgba(255, 220, 220, 0.2)',
     borderRadius: '12px',
-    padding: '16px',
-    marginBottom: '16px',
-    border: '1px solid rgba(255, 220, 220, 0.2)'
+    padding: '16px'
   };
 
   const buttonStyle = {
     width: '100%',
-    padding: '12px 16px', 
-    borderRadius: '8px', 
-    cursor: 'pointer',
+    padding: '10px 16px',
+    borderRadius: '8px',
+    border: 'none',
     fontSize: '14px',
     fontWeight: '600',
-    border: 'none',
-    textAlign: 'center',
+    cursor: 'pointer',
+    transition: 'all 0.2s ease',
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
-    gap: '8px',
-    marginBottom: '12px'
+    gap: '8px'
   };
 
   return (
@@ -292,6 +321,107 @@ const ProfilePage = ({ user, subscription, onLogout }) => {
           </div>
         </div>
 
+        {/* API Key Toggle Section */}
+        <div style={sectionStyle}>
+          <h4 style={{ 
+            color: '#FFDCDCFF', 
+            fontSize: '16px', 
+            fontWeight: '600', 
+            margin: '0 0 16px 0',
+            display: 'flex',
+            alignItems: 'center',
+            gap: '8px'
+          }}>
+            <FaKey />
+            API Configuration
+          </h4>
+
+          <div style={cardStyle}>
+            <div style={{ 
+              display: 'flex', 
+              alignItems: 'center', 
+              justifyContent: 'space-between',
+              gap: '12px'
+            }}>
+              <div style={{ flex: 1 }}>
+                <div style={{ 
+                  fontSize: '14px', 
+                  fontWeight: '600', 
+                  color: '#FFDCDCFF',
+                  marginBottom: '4px'
+                }}>
+                  Use Personal API Key
+                </div>
+                <div style={{ 
+                  fontSize: '12px', 
+                  color: 'rgba(255, 220, 220, 0.8)',
+                  lineHeight: '1.4'
+                }}>
+                  Turn on to use your own API key for unlimited usage
+                </div>
+              </div>
+              <button
+                onClick={handleTogglePersonalAPI}
+                disabled={isToggling}
+                style={{
+                  background: 'none',
+                  border: 'none',
+                  cursor: isToggling ? 'wait' : 'pointer',
+                  fontSize: '24px',
+                  color: subscription.userPreferPersonalAPI && subscription.hasPersonalKeys 
+                    ? '#17bf63' 
+                    : '#657786',
+                  display: 'flex',
+                  alignItems: 'center',
+                  padding: '4px',
+                  opacity: isToggling ? 0.6 : 1,
+                  transition: 'all 0.2s ease'
+                }}
+                title={
+                  subscription.userPreferPersonalAPI && subscription.hasPersonalKeys
+                    ? 'Click to use free trial API'
+                    : !subscription.hasPersonalKeys
+                      ? 'Configure API keys in settings first'
+                      : 'Click to use personal API key'
+                }
+              >
+                {subscription.userPreferPersonalAPI && subscription.hasPersonalKeys 
+                  ? <FaToggleOn /> 
+                  : <FaToggleOff />}
+              </button>
+            </div>
+            
+            {/* Status indicator */}
+            <div style={{ 
+              marginTop: '12px',
+              padding: '8px 12px',
+              borderRadius: '6px',
+              backgroundColor: subscription.usingPersonalAPI 
+                ? 'rgba(23, 191, 99, 0.1)' 
+                : subscription.remaining_requests <= 0 
+                  ? 'rgba(224, 36, 94, 0.1)'
+                  : 'rgba(255, 173, 31, 0.1)',
+              border: subscription.usingPersonalAPI 
+                ? '1px solid rgba(23, 191, 99, 0.3)' 
+                : subscription.remaining_requests <= 0 
+                  ? '1px solid rgba(224, 36, 94, 0.3)'
+                  : '1px solid rgba(255, 173, 31, 0.3)',
+              fontSize: '12px',
+              color: subscription.usingPersonalAPI 
+                ? '#17bf63' 
+                : subscription.remaining_requests <= 0 
+                  ? '#e0245e'
+                  : '#ffad1f'
+            }}>
+              {subscription.usingPersonalAPI 
+                ? '✅ Currently using your personal API key'
+                : subscription.hasPersonalKeys 
+                  ? '🔄 Currently using free trial API'
+                  : '⚠️ Configure API keys in settings to enable personal API'}
+            </div>
+          </div>
+        </div>
+
         {/* Subscription Status */}
         <div style={sectionStyle}>
           <h4 style={{ 
@@ -348,17 +478,17 @@ const ProfilePage = ({ user, subscription, onLogout }) => {
                 style={{
                   ...buttonStyle,
                   backgroundColor: subscription.remaining_requests <= 0 ? '#e0245e' : '#3b82f6',
-                  color: 'white',
-                  marginBottom: 0
+                  color: 'white'
                 }}
               >
-                {subscription.remaining_requests <= 0 ? 'Upgrade Now' : 'Manage Subscription'}
+                <FaChartBar />
+                {subscription.remaining_requests <= 0 ? 'Upgrade Plan' : 'Manage Subscription'}
               </button>
             )}
           </div>
         </div>
 
-        {/* Usage Statistics */}
+        {/* Usage Stats */}
         <div style={sectionStyle}>
           <h4 style={{ 
             color: '#FFDCDCFF', 
@@ -374,33 +504,15 @@ const ProfilePage = ({ user, subscription, onLogout }) => {
           </h4>
 
           <div style={cardStyle}>
-            {subscription.loading ? (
-              <div style={{ 
-                textAlign: 'center', 
-                color: 'rgba(255, 220, 220, 0.8)',
-                padding: '20px'
+            {subscription.usingPersonalAPI ? (
+              <div style={{
+                textAlign: 'center',
+                padding: '20px',
+                color: '#17bf63'
               }}>
-                Loading usage data...
-              </div>
-            ) : subscription.usingPersonalAPI ? (
-              <div style={{ textAlign: 'center', padding: '20px' }}>
-                <FaInfinity style={{ 
-                  fontSize: '32px', 
-                  color: '#17bf63',
-                  marginBottom: '8px'
-                }} />
-                <div style={{ 
-                  fontSize: '16px', 
-                  fontWeight: '600', 
-                  color: '#FFDCDCFF',
-                  marginBottom: '4px'
-                }}>
-                  Unlimited Usage
-                </div>
-                <div style={{ 
-                  fontSize: '12px', 
-                  color: 'rgba(255, 220, 220, 0.8)'
-                }}>
+                <FaInfinity style={{ fontSize: '32px', marginBottom: '8px' }} />
+                <div style={{ fontSize: '16px', fontWeight: '600' }}>Unlimited Usage</div>
+                <div style={{ fontSize: '12px', color: 'rgba(255, 220, 220, 0.8)' }}>
                   Using your personal API key
                 </div>
               </div>
@@ -409,60 +521,59 @@ const ProfilePage = ({ user, subscription, onLogout }) => {
                 <div style={{ 
                   display: 'flex', 
                   justifyContent: 'space-between', 
+                  alignItems: 'center',
                   marginBottom: '8px'
                 }}>
-                  <span style={{ 
-                    fontSize: '14px', 
-                    color: 'rgba(255, 220, 220, 0.8)'
-                  }}>
+                  <span style={{ fontSize: '14px', color: '#FFDCDCFF' }}>
                     Requests Used
                   </span>
                   <span style={{ 
                     fontSize: '14px', 
                     fontWeight: '600',
-                    color: '#FFDCDCFF'
+                    color: subscription.remaining_requests <= 0 ? '#e0245e' : '#FFDCDCFF'
                   }}>
                     {subscription.requests_used} / {subscription.monthly_request_limit}
                   </span>
                 </div>
-                
                 <div style={{
                   width: '100%',
                   height: '8px',
                   backgroundColor: 'rgba(255, 220, 220, 0.2)',
                   borderRadius: '4px',
-                  overflow: 'hidden',
-                  marginBottom: '12px'
+                  overflow: 'hidden'
                 }}>
                   <div style={{
-                    width: `${Math.min(100, (subscription.requests_used / subscription.monthly_request_limit) * 100)}%`,
+                    width: `${Math.min((subscription.requests_used / subscription.monthly_request_limit) * 100, 100)}%`,
                     height: '100%',
-                    backgroundColor: subscription.remaining_requests <= 2 ? '#e0245e' : '#17bf63',
+                    backgroundColor: subscription.remaining_requests <= 0 ? '#e0245e' : 
+                                   subscription.remaining_requests <= 2 ? '#ffad1f' : '#17bf63',
                     transition: 'width 0.3s ease'
                   }} />
                 </div>
-
                 <div style={{ 
                   fontSize: '12px', 
-                  color: subscription.remaining_requests <= 2 ? '#e0245e' : 'rgba(255, 220, 220, 0.8)',
+                  color: 'rgba(255, 220, 220, 0.8)',
+                  marginTop: '8px',
                   textAlign: 'center'
                 }}>
-                  {subscription.remaining_requests} requests remaining this month
+                  {subscription.remaining_requests > 0 
+                    ? `${subscription.remaining_requests} requests remaining`
+                    : 'Trial expired - upgrade or use personal API'
+                  }
                 </div>
               </div>
             )}
           </div>
         </div>
 
-        {/* Actions */}
+        {/* Logout Button */}
         <div style={sectionStyle}>
           <button
             onClick={onLogout}
             style={{
               ...buttonStyle,
               backgroundColor: '#e0245e',
-              color: 'white',
-              marginBottom: 0
+              color: 'white'
             }}
           >
             <FaSignOutAlt />
