+++ Social Shopping Agent/repo/src/components/SubscriptionChoice.js
