+++ Social Shopping Agent/repo/src/components/SubscriptionPage.js
