+++ Social Shopping Agent/repo/src/hooks/useAuth.js
