+++ extensions.json
@@ -121,22 +121,23 @@
       "updated_at": "2025-02-26T15:00:00Z"
     },
     {
-      "id": "mmdahpcmkbgemidfdlnpkmampbdijjai",
-      "name": "AI Chat Agent",
+      "id": "fkagldhhbbaillkjbjilcgflcihbkmoe",
+      "name": "Social Shopping Agent",
       "description": "An extension to assist users to chat and perform tasks using AI.",
       "version": "1.0.0",
       "author": "Wootzapp Team",
-      "download_url": "https://raw.githubusercontent.com/wootzapp/ext-store/main/AI%20Chat%20Agent/AIChatAgent.crx",  
-      "icon_url": "https://raw.githubusercontent.com/wootzapp/ext-store/main/AI%20Chat%20Agent/repo/public/icons/icon.png",
-      "github_url": "https://github.com/wootzapp/ext-store/tree/main/AI%20Chat%20Agent/repo",
+      "download_url": "https://raw.githubusercontent.com/wootzapp/ext-store/main/Social%20Shopping%20Agent/SocialShoppingAgent.crx",  
+      "icon_url": "https://raw.githubusercontent.com/wootzapp/ext-store/main/Social%20Shopping%20Agent/repo/public/icons/icon.png",
+      "github_url": "https://github.com/wootzapp/ext-store/tree/main/Social%20Shopping%20Agent/repo",
       "tags": [
-        "AI",
+        "AI Agent",
         "Social Media",
+        "Shopping Sites",
         "Productivity",
         "Automation"
       ],
-      "created_at": "2025-06-23T15:00:00Z",
-      "updated_at": "2025-06-23T15:00:00Z"
+      "created_at": "2025-08-01T15:00:00Z",
+      "updated_at": "2025-08-01T15:00:00Z"
     },
     {
       "id": "ficjckojhebnoogjmhacgodjcgddphac",
